Web additions
Have very preliminary visualizer for path of the crazyflie. At this point, it is not as close to 'real time' as desired but it does properly render the intended trajectory along the x axis and velocity on y. Will change along with the rest of the program to get position in x/y on axes in future, and see how the front/backend might be made more appropriately to render real-time data